requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sentence-transformers==3.0.0
faiss-cpu==1.13.2
numpy>=1.26.4
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
